refactor(server): moving shared module settings to common server folder
